Moved the helpers back to the same location as the class
