--- Plugins/RPGMenus/Source/RPGMenus/Public/Images/ImageAssetHelpers.h
@@ -1,43 +1,0 @@
-﻿// "Unreal Pokémon" created by Retro & Chill.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "ImageAsset.h"
-#include "Kismet/BlueprintFunctionLibrary.h"
-#include "ImageAssetHelpers.generated.h"
-
-
-/**
- * Blueprint helper functions for a variant type.
- */
-UCLASS()
-class RPGMENUS_API UImageAssetHelpers : public UBlueprintFunctionLibrary {
-    GENERATED_BODY()
-
-public:
-    UFUNCTION(BlueprintPure, DisplayName = "Make Image Asset (Texture)", Category = Variants, meta = (BlueprintAutocast))
-    static FImageAsset MakeImageAsset_Texture(UTexture* Texture);
-    
-    UFUNCTION(BlueprintPure, DisplayName = "Make Image Asset (Material)", Category = Variants, meta = (BlueprintAutocast))
-    static FImageAsset MakeImageAsset_Material(UMaterialInterface* Material);
-
-    UFUNCTION(BlueprintPure, DisplayName = "Make Image Asset (Slate Texture Atlas)", Category = Variants, meta = (BlueprintAutocast))
-    static FImageAsset MakeImageAsset_SlateTextureAsset(const TScriptInterface<ISlateTextureAtlasInterface>& TextureAtlas);
-
-    UFUNCTION(BlueprintPure, Category = Variants,
-              meta = (CompactNodeTitle = "->", BlueprintAutocast, AutoCreateRefTerm = ImageAsset))
-    static UObject* ConvertToObject(const FImageAsset& ImageAsset);
-
-    UFUNCTION(BlueprintCallable, Category = Variants,
-        meta = (AutoCreateRefTerm = "ImageAsset", ExpandEnumAsExecs = "ReturnValue"))
-    static EVariantFindResult CastToTexture(const FImageAsset& ImageAsset, UTexture*& AsTexture);
-
-    UFUNCTION(BlueprintCallable, Category = Variants,
-        meta = (AutoCreateRefTerm = "ImageAsset", ExpandEnumAsExecs = "ReturnValue"))
-    static EVariantFindResult CastToMaterial(const FImageAsset& ImageAsset, UMaterialInterface*& AsMaterial);
-
-    UFUNCTION(BlueprintCallable, Category = Variants,
-        meta = (AutoCreateRefTerm = "ImageAsset", ExpandEnumAsExecs = "ReturnValue"))
-    static EVariantFindResult CastToSlateTextureAtlas(const FImageAsset& ImageAsset, TScriptInterface<ISlateTextureAtlasInterface>& AsTextureAtlas);
-};
